+++ src/pages/styles.css
@@ -7,7 +7,20 @@
   padding: 40px;
 }
 
+.button {
+  font: bold 50px Arial;
+  text-decoration: none;
+  background-color: #EEEEEE;
+  color: #333333;
+  padding: 2px 6px 2px 6px;
+  border-top: 1px solid #CCCCCC;
+  border-right: 1px solid #333333;
+  border-bottom: 1px solid #333333;
+  border-left: 1px solid #CCCCCC;
+}
+
 .center {
+	text-align: center;
 	margin-left: auto;
 	margin-right: auto;
 	width:50%;
