--- src/src/styles.css
@@ -1,83 +1,0 @@
-.App {
-  font-family: sans-serif;
-  text-align: center;
-  display: grid;
-  grid-template-columns: 300px 1fr;
-  column-gap: 32px;
-  padding: 40px;
-}
-
-.responsive {
-  max-width: 100%;
-  height: auto;
-}
-
-.tip {
-  color: #777;
-  margin-top: -16px;
-  margin-bottom: 32px;
-}
-
-.rooms {
-  text-align: left;
-  font-size: 18px;
-}
-
-.square {
-  display: inline-block;
-  height: 15px;
-  width: 15px;
-  background-color: blue;
-  margin-right: 10px;
-  margin-bottom: -1px;
-  border-radius: 3px;
-}
-
-.room-link {
-  cursor: pointer;
-  border: 1px solid #e1e1e1;
-  padding: 12px 12px;
-  margin-top: 4px;
-  border-radius: 2px;
-}
-
-.room-link.active {
-  border-color: #333;
-  font-weight: 600;
-}
-
-.room {
-  fill-opacity: 0.1;
-  stroke-opacity: 0;
-  stroke-width: 2;
-}
-
-.room.active {
-  fill-opacity: 0.2;
-  stroke-opacity: 1;
-}
-
-.kitchen {
-  fill: #00ced1;
-  stroke: #00ced1;
-}
-
-.hissit {
-  fill: #3b82f6;
-  stroke: #3b82f6;
-}
-
-.wc {
-  fill: #ff1493;
-  stroke: #ff1493;
-}
-
-.auditorio {
-  fill: #a229b6;
-  stroke: #a229b6;
-}
-
-.ulysseus {
-	fill: #34d399;
-	stroke: #34d399;
-}
